Break init_db code into multiple files
