+++ lambdas/setup_db/src/utils.py
@@ -1,0 +1,7 @@
+import os
+
+
+def get_environ_params(*args):
+    return [
+        os.environ[k] for k in args
+    ]
